Model: Move the database handling to a helper class
To make it simpler to handle the database authentification at the
application level move the database access out of the model. This is
done with a new model class that forward all the low level methods
to a database connection object. This object also provide methods to
handle login and logout for the application.