Cleanup the app structure
The app must now be provided a DB object at init time. The app then
use this DB object to create all the model and view/control classes.

The app now also register a callback on the DB object to allow it
to query the user credentials. This make sure that the user can always
re-login when the session timeout.