Load the EJS templates with steal to avoid caching problems
When the EJS templates are loaded separatly the cache busting is not
working properly when we update the application. This lead to strange
bugs as the templates are not matching the code anymore. The simplest
way is to just let steal load the templates with the rest of the code.