coplanar/calendar: Allow sub class to extend the UI with templates
The sub classes can use a templates to add some custom UI around the
calendar view. For example custom navigation links, or some filtering
selection.

The template must call the setCalendarElement() function on the HTML
element that should become the calendar view. The template
environement also provide the bindRouteAttr() funtion to bind a form
input element with a route attribute. The element will then be updated
when the route change, and it will set the route when edited.